Add an optimized PutImage using CPU blit
Modern applications often write into an off-screen pixmap or
image that is regularly transferred to the screen window.
Depending on the method used, it appears that the code paths
used by the Xorg fb driver for this purpose may not always be
optimal. This is especially true if the ShmPutImage method is
used, and to lesser extent for normal (piped) PutImage. The
newer shared memory pixmap type is not affected and seems to
be already optimized.

Looking a the Xorg ShmPutImage implementation, it may use
either CopyArea (which is optimized in the sunxifb driver) or
PutImage. Running x11perf -shmput500 triggers the CopyArea
path so it does not show the weakness. However, it appears
that the PutImage path is regularly used by applications in
practice.

This implementation does not use G2D, so should be usable on
different platforms. It shows a clear improvement in certain
off-screen to screen blits by using pixman instead of the
default Xorg fb implementation.

Using a slightly modified version
(https://github.com/hglm/patches/benchimagemark.tar.gz) of
a program I found on the web that repeatedly writes into an
image and transfers it to the screen, here are some benchmarks:

benchimagemark512 at 1920x0180x32bppx60Hz:

Before:

1: XPutImage(XImage): 34.548875755 (14.47 Mpix/s)
2: XPutImage(XImage): Pixmap, XCopyArea(Pixmap): 37.251380690 (13.42 Mpix/s)
3: XPutImage(XImage): Pixmap, XRenderComposite(Pixmap): 37.170898615 (13.45 Mpix/s)
4: XShmPutImage(XImage SHM): 18.167374795 (27.52 Mpix/s)
5: XShmPutImage(XImage SHM): Pixmap, XCopyArea(Pixmap): 23.045493400 (21.70 Mpix/s)
6: XShmPutImage(XImage SHM): Pixmap, XRenderComposite(Pixmap): 23.069486572 (21.67 Mpix/s)
7: XCopyArea(Pixmap SHM): 12.920413918 (38.70 Mpix/s)
8: XRenderComposite(Pixmap SHM): 12.972882531 (38.54 Mpix/s)

After:

1: XPutImage(XImage): 29.952659937 (16.69 Mpix/s)
2: XPutImage(XImage): Pixmap, XCopyArea(Pixmap): 33.122204665 (15.10 Mpix/s)
3: XPutImage(XImage): Pixmap, XRenderComposite(Pixmap): 32.846011813 (15.22 Mpix/s)
4: XShmPutImage(XImage SHM): 13.359728942 (37.43 Mpix/s)
5: XShmPutImage(XImage SHM): Pixmap, XCopyArea(Pixmap): 17.665096856 (28.30 Mpix/s)
6: XShmPutImage(XImage SHM): Pixmap, XRenderComposite(Pixmap): 17.660086501 (28.31 Mpix/s)
7: XCopyArea(Pixmap SHM): 13.310822346 (37.56 Mpix/s)
8: XRenderComposite(Pixmap SHM): 13.360211359 (37.42 Mpix/s)

At lower resolutions and lower color depth the improvement is
greater. The improvement also get bigger when benchimagemark is
configured not to fill the off-screen area each iteration. For
example, at 1280x720x16bppx60Hz ShmPutImage throughput improves
from 39 to 70 Mpix/s in this configuration.