Add Fill function to device-independent 2D acceleration interface
This patch adds a fill function to the device-independent blt2d_i
structure in interfaces.h. In the CPU backend, the fill function
is initialized with noop. In the G2D backend, the fill function
is initialized with the G2D fill function. In the sunxi G2D backend
(sunxi_disp.c), add a G2D fill function that supports both 16bpp
and 32bpp. In the X driver (sunxi_x_g2d.c), the private fill
function is initialized with the blt2d fill function. It not used
yet.