Change thresholds for falling back from G2D to CPU blit/fill
When G2D is enabled, using G2D for small area blits and fills
incurs an overhead that makes it slower than the CPU version. The
area thresholds below which the CPU is used are now as follows:

32bpp screen-to-screeen blit: 1000
16bpp screen-to-screen blit: 2500
32bpp rectangle fill: 5000
16bpp rectangle fill: 10000000 (only CPU fill used in practice)

For 32bpp fill, the current settings makes fills greater than 5000
pixels in size a bit slower because G2D is slower than CPU fill.
On the other hand, CPU utilization is much lower with G2D.

With the current setting, 16bpp G2D fill is practically disabled;
the slow down when using G2D is too great (CPU fill is significantly
faster).