Use pixman as fallback in xCopyWindowProc if possible
Since it is known that fbBlt is slow, it is useful to avoid it if
pixman can be used instead. When ShadowFB is enabled, dragging
certain windows (e.g. Leafpad and Midori but not lxterminal in LXDE)
generates lots of CopyWindow requests with a source that is not in
the uncachable framebuffer. In this case, the overlapped_blt
function returns false and fbBlt is used instead. By using pixman
if possible (reverse and upsidedown flags not set), a speed-up is
obtained.

Since the ShadowFB is disabled by default on most hardware in the
most recent versions of sunxifb, this change does not have a great
practical benefit. Still, it should help in configurations were
ShadowFB is enabled by default or when ShadowFB is enabled for some
reason.

Signed-off-by: Harm Hanemaaijer <fgenfb@yahoo.com>